+++ libs/schematics/src/generators/admin-app/schema.d.ts
@@ -1,7 +1,11 @@
-export interface AdminAppGeneratorSchema extends Record<string, unknown> {
+export interface AdminAppGeneratorSchema {
   name: string;
   directory?: string;
   npmName?: string;
   reactAdminImportPath: string;
   rextClientImportPath: string;
 }
+
+export interface AdminAppGeneratorSchemaWithExtra
+  extends AdminAppGeneratorSchema,
+    Record<string, unknown> {}