+++ libs/schematics/src/generators/traxion-workspace/helpers/add-npmrc.helper.ts
@@ -1,0 +1,5 @@
+import { Tree } from '@nrwl/devkit';
+
+export function addNpmrc(tree: Tree) {
+  tree.write('.npmrc', '@tractr:registry=https://npm.pkg.github.com');
+}