+++ libs/schematics/src/helpers/get-import-prefix-path.ts
@@ -1,0 +1,6 @@
+import { getWorkspaceLayout, Tree } from '@nrwl/devkit';
+
+export function getImportPrefixPath(tree: Tree, directory = '') {
+  const { npmScope } = getWorkspaceLayout(tree);
+  return `@${npmScope}/${directory ? `${directory}-` : ''}`;
+}