+++ libs/schematics/src/schematics.constants.ts
@@ -1,3 +1,8 @@
+import { TargetConfiguration } from '@nrwl/devkit';
+
+import { AvailableTractrTemplates } from './generators/hapify-library/schema';
+import { Packages, PackageType } from './helpers';
+
 export const DEFAULT_IMPORT_REPLACEMENTS = {
   'angular-rext-client': ['rext-client'],
   casl: [],
@@ -6,7 +11,7 @@ export const DEFAULT_IMPORT_REPLACEMENTS = {
   'nestjs-models': ['nestjs-models-common', 'nestjs-models-rest'],
   'nestjs-models-common': ['models'],
   'nestjs-models-rest': ['casl', 'nestjs-models-common', 'rest-dtos', 'models'],
-  prisma: [],
+  prisma: ['nestjs-models', 'nestjs-models-common'],
   'react-admin': ['rext-client'],
   'rest-dtos': ['models'],
   'rext-client': ['models', 'rest-dtos'],
@@ -24,4 +29,70 @@ export const DEFAULT_SECONDARY_ENTRY_POINTS: Record<string, string[]> = {
   'react-admin': [],
   'rest-dtos': ['mock'],
   'rext-client': [],
+};
+
+export const DEFAULT_LIBRARY_TYPE: Record<AvailableTractrTemplates, string> = {
+  'angular-rext-client': 'angular',
+  casl: 'nest',
+  dbml: 'nest',
+  models: 'nest',
+  'nestjs-models': 'nest',
+  'nestjs-models-common': 'nest',
+  'nestjs-models-rest': 'nest',
+  prisma: 'nest',
+  'react-admin': 'react',
+  'rest-dtos': 'nest',
+  'rext-client': 'angular',
+};
+
+export type LibraryUseContext = 'angular' | 'nest' | 'react';
+export const DEFAULT_LIBRARY_USE_CONTEXT: Record<
+  AvailableTractrTemplates,
+  LibraryUseContext[]
+> = {
+  'angular-rext-client': ['angular'],
+  casl: ['nest', 'angular', 'react'],
+  dbml: ['nest'],
+  models: ['nest', 'angular', 'react'],
+  'nestjs-models': ['nest'],
+  'nestjs-models-common': ['nest'],
+  'nestjs-models-rest': ['nest'],
+  prisma: ['nest'],
+  'react-admin': ['react'],
+  'rest-dtos': ['nest', 'angular', 'react'],
+  'rext-client': ['nest', 'angular', 'react'],
+};
+
+export const DEFAULT_TARGETS_OPTIONS: Partial<
+  Record<
+    AvailableTractrTemplates,
+    Record<string, Partial<TargetConfiguration> | null>
+  >
+> = {
+  dbml: {
+    build: null,
+    lint: null,
+    generate: {
+      options: {
+        format: false,
+        moveGeneratedFiles: false,
+        updateImportPath: false,
+      },
+    },
+  },
+};
+
+export const DEFAULT_DEPENDENCIES: Partial<
+  Record<AvailableTractrTemplates, Packages>
+> = {
+  models: [
+    { packageName: 'faker', version: '5.5.3', type: PackageType.dependencies },
+  ],
+  'angular-rext-client': [
+    {
+      packageName: '@tractr/angular-tools',
+      version: 'current',
+      type: PackageType.dependencies,
+    },
+  ],
 };