+++ libs/terraform/helpers/post-synth/src/post-synth.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+import { fixCidrSubnet } from './fix-cidrsubnet';
+
+fixCidrSubnet();