+++ libs/CHANGELOG.md
@@ -2,6 +2,204 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.39.4](https://github.com/tractr/stack/compare/v1.39.3...v1.39.4) (2022-02-22)
+
+
+### Bug Fixes
+
+* update create traxion package name ([#395](https://github.com/tractr/stack/issues/395)) ([2a3013e](https://github.com/tractr/stack/commit/2a3013ef4e2f88f6232e1fa524ed01b18306be8e))
+
+
+
+## [1.39.4](https://github.com/tractr/stack/compare/v1.39.3...v1.39.4) (2022-02-22)
+
+
+### Bug Fixes
+
+* update create traxion package name ([#395](https://github.com/tractr/stack/issues/395)) ([2a3013e](https://github.com/tractr/stack/commit/2a3013ef4e2f88f6232e1fa524ed01b18306be8e))
+
+
+
+## [1.39.3](https://github.com/tractr/stack/compare/v1.39.2...v1.39.3) (2022-02-22)
+
+
+### Bug Fixes
+
+* update authentication module and templates to fix public routes ([#391](https://github.com/tractr/stack/issues/391)) ([3f72027](https://github.com/tractr/stack/commit/3f72027e10b4a480fd5d8b462f19ea04216cc69d))
+
+
+
+## [1.39.3](https://github.com/tractr/stack/compare/v1.39.2...v1.39.3) (2022-02-22)
+
+
+### Bug Fixes
+
+* update authentication module and templates to fix public routes ([#391](https://github.com/tractr/stack/issues/391)) ([3f72027](https://github.com/tractr/stack/commit/3f72027e10b4a480fd5d8b462f19ea04216cc69d))
+
+
+
+## [1.39.2](https://github.com/tractr/stack/compare/v1.39.1...v1.39.2) (2022-02-22)
+
+
+### Bug Fixes
+
+* make the hapify workspace schematics work ([#390](https://github.com/tractr/stack/issues/390)) ([01231cb](https://github.com/tractr/stack/commit/01231cb1790dc196504a27374e5552907245c303))
+
+
+
+## [1.39.2](https://github.com/tractr/stack/compare/v1.39.1...v1.39.2) (2022-02-22)
+
+
+### Bug Fixes
+
+* make the hapify workspace schematics work ([#390](https://github.com/tractr/stack/issues/390)) ([01231cb](https://github.com/tractr/stack/commit/01231cb1790dc196504a27374e5552907245c303))
+
+
+
+## [1.39.1](https://github.com/tractr/stack/compare/v1.39.0...v1.39.1) (2022-02-19)
+
+
+### Bug Fixes
+
+* **terraform-helpers-post-synth:** fix terraform output path detection ([#393](https://github.com/tractr/stack/issues/393)) ([8a2f07d](https://github.com/tractr/stack/commit/8a2f07dc9fd7d4475d02f5627bdd1228b41501ad))
+
+
+
+## [1.39.1](https://github.com/tractr/stack/compare/v1.39.0...v1.39.1) (2022-02-19)
+
+
+### Bug Fixes
+
+* **terraform-helpers-post-synth:** fix terraform output path detection ([#393](https://github.com/tractr/stack/issues/393)) ([8a2f07d](https://github.com/tractr/stack/commit/8a2f07dc9fd7d4475d02f5627bdd1228b41501ad))
+
+
+
+# [1.39.0](https://github.com/tractr/stack/compare/v1.38.0...v1.39.0) (2022-02-18)
+
+
+### Features
+
+* **terraform-helpers-post-synth:** fix publish config ([#392](https://github.com/tractr/stack/issues/392)) ([3a0eac7](https://github.com/tractr/stack/commit/3a0eac7f9bccfa2759974c1474e1daa98c2e7149))
+
+
+
+# [1.39.0](https://github.com/tractr/stack/compare/v1.38.0...v1.39.0) (2022-02-18)
+
+
+### Features
+
+* **terraform-helpers-post-synth:** fix publish config ([#392](https://github.com/tractr/stack/issues/392)) ([3a0eac7](https://github.com/tractr/stack/commit/3a0eac7f9bccfa2759974c1474e1daa98c2e7149))
+
+
+
+# [1.38.0](https://github.com/tractr/stack/compare/v1.37.1...v1.38.0) (2022-02-17)
+
+
+### Features
+
+* add create hapify workspace command ([#387](https://github.com/tractr/stack/issues/387)) ([98119b9](https://github.com/tractr/stack/commit/98119b9be282b5afcd8bf7ee82af8b677eeaab0a))
+
+
+
+# [1.38.0](https://github.com/tractr/stack/compare/v1.37.1...v1.38.0) (2022-02-17)
+
+
+### Features
+
+* add create hapify workspace command ([#387](https://github.com/tractr/stack/issues/387)) ([98119b9](https://github.com/tractr/stack/commit/98119b9be282b5afcd8bf7ee82af8b677eeaab0a))
+
+
+
+## [1.37.1](https://github.com/tractr/stack/compare/v1.37.0...v1.37.1) (2022-02-17)
+
+
+### Bug Fixes
+
+* update type for nests-common password middleware ([#389](https://github.com/tractr/stack/issues/389)) ([c9a4237](https://github.com/tractr/stack/commit/c9a42374d825815b0ad470c05ab3c3425aca3b03))
+
+
+
+## [1.37.1](https://github.com/tractr/stack/compare/v1.37.0...v1.37.1) (2022-02-17)
+
+
+### Bug Fixes
+
+* update type for nests-common password middleware ([#389](https://github.com/tractr/stack/issues/389)) ([c9a4237](https://github.com/tractr/stack/commit/c9a42374d825815b0ad470c05ab3c3425aca3b03))
+
+
+
+# [1.37.0](https://github.com/tractr/stack/compare/v1.36.3...v1.37.0) (2022-02-16)
+
+
+### Features
+
+* add prisma schematic ([#388](https://github.com/tractr/stack/issues/388)) ([f68d477](https://github.com/tractr/stack/commit/f68d4772c78a18e9b8c20f3908e25cdc11a5a6b2))
+
+
+
+# [1.37.0](https://github.com/tractr/stack/compare/v1.36.3...v1.37.0) (2022-02-16)
+
+
+### Features
+
+* add prisma schematic ([#388](https://github.com/tractr/stack/issues/388)) ([f68d477](https://github.com/tractr/stack/commit/f68d4772c78a18e9b8c20f3908e25cdc11a5a6b2))
+
+
+
+## [1.36.3](https://github.com/tractr/stack/compare/v1.36.2...v1.36.3) (2022-02-11)
+
+
+### Bug Fixes
+
+* **terraform-group-network:** add post-synth script ([#384](https://github.com/tractr/stack/issues/384)) ([9b25040](https://github.com/tractr/stack/commit/9b250405074adb4cf53e64ea61cb24b837ff8c54))
+
+
+
+## [1.36.3](https://github.com/tractr/stack/compare/v1.36.2...v1.36.3) (2022-02-11)
+
+
+### Bug Fixes
+
+* **terraform-group-network:** add post-synth script ([#384](https://github.com/tractr/stack/issues/384)) ([9b25040](https://github.com/tractr/stack/commit/9b250405074adb4cf53e64ea61cb24b837ff8c54))
+
+
+
+## [1.36.2](https://github.com/tractr/stack/compare/v1.36.1...v1.36.2) (2022-02-10)
+
+
+### Bug Fixes
+
+* add modules to imports properties when using @Module ([#385](https://github.com/tractr/stack/issues/385)) ([72b381b](https://github.com/tractr/stack/commit/72b381b6a5cb5b985d11fb4f1bc260351243ba99))
+
+
+
+## [1.36.2](https://github.com/tractr/stack/compare/v1.36.1...v1.36.2) (2022-02-10)
+
+
+### Bug Fixes
+
+* add modules to imports properties when using @Module ([#385](https://github.com/tractr/stack/issues/385)) ([72b381b](https://github.com/tractr/stack/commit/72b381b6a5cb5b985d11fb4f1bc260351243ba99))
+
+
+
+## [1.36.1](https://github.com/tractr/stack/compare/v1.36.0...v1.36.1) (2022-02-09)
+
+
+### Bug Fixes
+
+* **terraform-group-zone:** remove deprecated workaround ([#383](https://github.com/tractr/stack/issues/383)) ([59e0196](https://github.com/tractr/stack/commit/59e0196da963d559534d9453614c93e7c21a8b19))
+
+
+
+## [1.36.1](https://github.com/tractr/stack/compare/v1.36.0...v1.36.1) (2022-02-09)
+
+
+### Bug Fixes
+
+* **terraform-group-zone:** remove deprecated workaround ([#383](https://github.com/tractr/stack/issues/383)) ([59e0196](https://github.com/tractr/stack/commit/59e0196da963d559534d9453614c93e7c21a8b19))
+
+
+
 # [1.36.0](https://github.com/tractr/stack/compare/v1.35.0...v1.36.0) (2022-02-08)