+++ libs/hapify/templates/nestjs-models/CHANGELOG.md
@@ -2,6 +2,60 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.39.4](https://github.com/tractr/stack/compare/v1.39.3...v1.39.4) (2022-02-22)
+
+
+
+## [1.39.3](https://github.com/tractr/stack/compare/v1.39.2...v1.39.3) (2022-02-22)
+
+
+
+## [1.39.2](https://github.com/tractr/stack/compare/v1.39.1...v1.39.2) (2022-02-22)
+
+
+
+## [1.39.1](https://github.com/tractr/stack/compare/v1.39.0...v1.39.1) (2022-02-19)
+
+
+
+# [1.39.0](https://github.com/tractr/stack/compare/v1.38.0...v1.39.0) (2022-02-18)
+
+
+
+# [1.38.0](https://github.com/tractr/stack/compare/v1.37.1...v1.38.0) (2022-02-17)
+
+
+### Features
+
+* add create hapify workspace command ([#387](https://github.com/tractr/stack/issues/387)) ([98119b9](https://github.com/tractr/stack/commit/98119b9be282b5afcd8bf7ee82af8b677eeaab0a))
+
+
+
+## [1.37.1](https://github.com/tractr/stack/compare/v1.37.0...v1.37.1) (2022-02-17)
+
+
+
+# [1.37.0](https://github.com/tractr/stack/compare/v1.36.3...v1.37.0) (2022-02-16)
+
+
+
+## [1.36.3](https://github.com/tractr/stack/compare/v1.36.2...v1.36.3) (2022-02-11)
+
+
+
+## [1.36.2](https://github.com/tractr/stack/compare/v1.36.1...v1.36.2) (2022-02-10)
+
+
+### Bug Fixes
+
+* add modules to imports properties when using @Module ([#385](https://github.com/tractr/stack/issues/385)) ([72b381b](https://github.com/tractr/stack/commit/72b381b6a5cb5b985d11fb4f1bc260351243ba99))
+
+
+
+## [1.36.1](https://github.com/tractr/stack/compare/v1.36.0...v1.36.1) (2022-02-09)
+
+
+
 # [1.36.0](https://github.com/tractr/stack/compare/v1.35.0...v1.36.0) (2022-02-08)