+++ libs/schematics/CHANGELOG.md
@@ -2,6 +2,65 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.39.4](https://github.com/tractr/stack/compare/v1.39.3...v1.39.4) (2022-02-22)
+
+
+
+## [1.39.3](https://github.com/tractr/stack/compare/v1.39.2...v1.39.3) (2022-02-22)
+
+
+
+## [1.39.2](https://github.com/tractr/stack/compare/v1.39.1...v1.39.2) (2022-02-22)
+
+
+### Bug Fixes
+
+* make the hapify workspace schematics work ([#390](https://github.com/tractr/stack/issues/390)) ([01231cb](https://github.com/tractr/stack/commit/01231cb1790dc196504a27374e5552907245c303))
+
+
+
+## [1.39.1](https://github.com/tractr/stack/compare/v1.39.0...v1.39.1) (2022-02-19)
+
+
+
+# [1.39.0](https://github.com/tractr/stack/compare/v1.38.0...v1.39.0) (2022-02-18)
+
+
+
+# [1.38.0](https://github.com/tractr/stack/compare/v1.37.1...v1.38.0) (2022-02-17)
+
+
+### Features
+
+* add create hapify workspace command ([#387](https://github.com/tractr/stack/issues/387)) ([98119b9](https://github.com/tractr/stack/commit/98119b9be282b5afcd8bf7ee82af8b677eeaab0a))
+
+
+
+## [1.37.1](https://github.com/tractr/stack/compare/v1.37.0...v1.37.1) (2022-02-17)
+
+
+
+# [1.37.0](https://github.com/tractr/stack/compare/v1.36.3...v1.37.0) (2022-02-16)
+
+
+### Features
+
+* add prisma schematic ([#388](https://github.com/tractr/stack/issues/388)) ([f68d477](https://github.com/tractr/stack/commit/f68d4772c78a18e9b8c20f3908e25cdc11a5a6b2))
+
+
+
+## [1.36.3](https://github.com/tractr/stack/compare/v1.36.2...v1.36.3) (2022-02-11)
+
+
+
+## [1.36.2](https://github.com/tractr/stack/compare/v1.36.1...v1.36.2) (2022-02-10)
+
+
+
+## [1.36.1](https://github.com/tractr/stack/compare/v1.36.0...v1.36.1) (2022-02-09)
+
+
+
 # [1.36.0](https://github.com/tractr/stack/compare/v1.35.0...v1.36.0) (2022-02-08)